learn plugin: Handle channels with custom presentation
When a channel is selected, check if there is a node with a custom
channel presentation for the channel. By querying if there is a node
with a convention in its metadata.

The current convention is:
- kind is "html5", and
- title is "custom-channel-ui"

If there is such node, navigate to it directly. If not, show the usual
presentation.

https://phabricator.endlessm.com/T30628